% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_big.R
\name{as.hash_big.matrix}
\alias{as.hash_big.matrix}
\alias{attach.hash_big.matrix}
\title{The core "hash_big.matrix" operations}
\usage{
as.hash_big.matrix(x, backingpath = "bp", silence = TRUE, ...)

attach.hash_big.matrix(x, backingpath = "bp")
}
\arguments{
\item{x}{A matrix, vector, or data.frame for \code{\link[bigmemory]{as.big.matrix}}.}

\item{backingpath, ...}{Passed to \code{\link[bigmemory]{as.big.matrix}}.}

\item{silence}{Suppress messages?}
}
\description{
The core "hash_big.matrix" operations
}
