% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{attach_all_matrices}
\alias{attach_all_matrices}
\title{Attach all matrices in a batch simulation}
\usage{
attach_all_matrices(bs, backingpath = "bp")
}
\arguments{
\item{bs}{A \code{batch_simulation} object.}

\item{backingpath}{Passed to \code{\link[bigmemory:as.big.matrix]{bigmemory::as.big.matrix()}}.}
}
\value{
A \code{batch_simulation} object with all \code{hash_big_matrix}es attached.
}
\description{
Attach all matrices in a batch simulation
}
