% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{autolayer.barrier}
\alias{autolayer.barrier}
\title{Get a ggplot2 layer from a barrier object}
\usage{
\method{autolayer}{barrier}(object, path = TRUE, ...)
}
\arguments{
\item{object}{A \code{barrier} object.}

\item{path}{Show the minimum energy path in the graph?}

\item{...}{Not in use.}
}
\value{
A \code{ggplot2} layer that can be added to an existing landscape plot.
}
\description{
This layer can show the saddle point (2d) and the minimal energy path (3d) on the landscape.
}
