% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_landscape.R
\name{autoplot.landscape}
\alias{autoplot.landscape}
\title{Autoplot landscape objects}
\usage{
\method{autoplot}{landscape}(object, ...)
}
\arguments{
\item{object}{A landscape object.}

\item{...}{Not in use.}
}
\value{
A ggplot object.
}
\description{
Returns the ggplot representation stored in a landscape object.
}
