% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{calculate_barrier_2d}
\alias{calculate_barrier_2d}
\alias{get_barrier_height}
\title{Calculate barrier from a 2D distribution}
\usage{
calculate_barrier_2d(
  d,
  start_location = c(0, 0),
  start_r = 0.1,
  end_location = c(0.7, 0.6),
  end_r = 0.15
)

get_barrier_height(b)
}
\arguments{
\item{d}{A \code{3d_static_landscape} object or a \code{kde2d} distribution.}

\item{start_location, end_location}{The initial position for searching the start/end point.}

\item{start_r, end_r}{The searching radius for searching the start/end point.}

\item{b}{A \code{barrier_landscape} object.}
}
\description{
Calculate barrier from a 2D distribution
}
\section{Functions}{
\itemize{
\item \code{get_barrier_height}: Get the barrier height.
}}

