% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier_batch.R
\name{calculate_barrier_2d_batch}
\alias{calculate_barrier_2d_batch}
\title{Calculate barrier from a 2D landscape with multiple simulations}
\usage{
calculate_barrier_2d_batch(
  l,
  bg = NULL,
  start_location_value = 0,
  start_r = 0.1,
  end_location_value = 0.7,
  end_r = 0.15,
  base = exp(1)
)
}
\arguments{
\item{l}{A \code{2d_animation_landscape} (not implemented yet) or a \code{2d_matrix_landscape}.}

\item{bg}{A \code{barrier_grid_3d} object if you want to use different parameters for each condition. Otherwise \code{NULL}.}

\item{start_location_value, end_location_value}{The initial position (in value) for searching the start/end point.}

\item{start_r, end_r}{The searching (L1) radius for searching the start/end point.}

\item{base}{The base of the log function.}
}
\description{
Calculate barrier from a 2D landscape with multiple simulations
}
