% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{calculate_barrier_3d_batch}
\alias{calculate_barrier_3d_batch}
\title{Calculate barrier from a 3D landscape with multiple simulations}
\usage{
calculate_barrier_3d_batch(
  l,
  start_location_value = c(0, 0),
  start_r = 0.1,
  end_location_value = c(0.7, 0.6),
  end_r = 0.15,
  base = exp(1)
)
}
\arguments{
\item{l}{A \code{3d_animation_landscape} or a \code{3d_matrix_landscape}.}

\item{start_location_value, end_location_value}{The initial position (in value) for searching the start/end point.}

\item{start_r, end_r}{The searching (L1) radius for searching the start/end point.}

\item{base}{The base of the log function.}
}
\description{
Calculate barrier from a 3D landscape with multiple simulations
}
