% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{calculate_barrier_3d_core}
\alias{calculate_barrier_3d_core}
\title{Shared 3D barrier core}
\usage{
calculate_barrier_3d_core(
  d,
  x_label,
  y_label,
  start_location_value,
  start_r,
  end_location_value,
  end_r,
  Umax,
  expand = TRUE,
  omit_unstable = FALSE,
  base = exp(1)
)
}
\arguments{
\item{d}{A distribution object with components \code{x}, \code{y}, and matrix \code{d}.}

\item{x_label, y_label}{Axis labels for plotting.}

\item{start_location_value, end_location_value}{The initial position (in value)
for searching the start/end point.}

\item{start_r, end_r}{The search radius (in L1 distance) for the start/end point.}

\item{Umax}{The highest possible value of the potential function.}

\item{expand}{If the values in the range all equal to \code{Umax}, expand the
range or not?}

\item{omit_unstable}{If a state is not stable (the "local minimum" overlaps
with the saddle point), omit that state or not?}

\item{base}{The base of the log function.}
}
\value{
A \verb{3d_barrier} object.
}
\description{
Internal helper used by \code{simlandr} and downstream packages to calculate barrier heights.
}
\keyword{internal}
