% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{fill_in_struct}
\alias{fill_in_struct}
\title{Fill a vector of values into a structure list.}
\usage{
fill_in_struct(vec, struct)
}
\arguments{
\item{vec}{A vector of values.}

\item{struct}{A list with a certain structure.}
}
\value{
A \code{ele_list} object.
}
\description{
Fill a vector of values into a structure list.
}
\seealso{
\code{\link[=modified_simulation]{modified_simulation()}}
}
\keyword{internal}
