% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{find_local_min}
\alias{find_local_min}
\title{Find local minimum of a distribution}
\usage{
find_local_min(dist, localmin, r)
}
\arguments{
\item{dist}{An \code{kde2d} distribution object.}

\item{localmin}{Starting value of finding local minimum}

\item{r}{Searching radius}
}
\description{
Find local minimum of a distribution
}
