% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{get_barrier_height}
\alias{get_barrier_height}
\title{Summarize the barrier height from a \code{barrier} object}
\usage{
get_barrier_height(object, ...)
}
\arguments{
\item{object}{A \code{barrier} object.}

\item{...}{Not in use.}
}
\value{
A vector (for a single barrier calculation result) or a \code{data.frame} (for batch barrier calculation results) that contains the barrier heights on the landscape.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated. Use \code{\link[=summary]{summary()}} instead. (See \code{\link[=summary.barrier]{summary.barrier()}}.)
}
\keyword{internal}
