% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_landscape.R
\name{get_dist}
\alias{get_dist}
\title{Get the probability distribution from a landscape object}
\usage{
get_dist(l, index = 1)
}
\arguments{
\item{l}{A \code{landscape} project.}

\item{index}{1 to get the distribution in tidy format; 2 or "raw" to get the raw simulation result (\code{batch_simulation}).}
}
\description{
Get the probability distribution from a landscape object
}
