% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_big.R
\docType{class}
\name{hash_big.matrix-class}
\alias{hash_big.matrix-class}
\title{Big matrix with a md5 hash reference}
\description{
Big matrix with a md5 hash reference
}
\section{Slots}{

\describe{
\item{\code{md5}}{The md5 value of the matrix.}

\item{\code{address}}{Inherited from \code{big.matrix}.}
}}

