% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_api.R
\name{make_2d_tidy_dist}
\alias{make_2d_tidy_dist}
\title{Make a tidy \code{data.frame} from smooth 2d distribution matrix}
\usage{
make_2d_tidy_dist(dist_2d, value = NULL, var_name = NULL)
}
\arguments{
\item{dist_2d}{\code{kde2d} distribution.}

\item{value}{The value of the variable of interest.}

\item{var_name}{The name of the variable.}
}
\value{
A tidy \code{data.frame}.
}
\description{
Make a tidy \code{data.frame} from smooth 2d distribution matrix
}
\keyword{internal}
