% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_simulation_landscape.R
\name{make_3d_kernel_dist}
\alias{make_3d_kernel_dist}
\title{Make 3D kernel smooth distribution}
\usage{
make_3d_kernel_dist(
  output,
  x,
  y,
  z,
  n = 200,
  lims = c(-0.1, 1.1, -0.1, 1.1, -0.1, 1.1),
  h
)
}
\arguments{
\item{output}{A matrix of simulation output.}

\item{x, y, z}{The name of the target variable.}

\item{n, lims, h}{Passed to \code{\link[ks]{kde}} (but using the format of \code{\link[MASS]{kde2d}} to make it consistent across functions).
For \code{ks::kde}, \code{H = diag(h, 2, 2)}.}
}
\value{
A \code{MASS::kde2d}-type list.
}
\description{
Make 3D kernel smooth distribution
}
