% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_simulation_landscape.R
\name{make_3d_tidy_dist}
\alias{make_3d_tidy_dist}
\title{Make a tidy data frame from smooth 3d distribution matrix}
\usage{
make_3d_tidy_dist(dist_3d, value = NULL, var_name = NULL)
}
\arguments{
\item{dist_3d}{\code{kde2d}-type distribution.}

\item{value}{The value of the variable of interest.}

\item{var_name}{The name of the variable.}
}
\value{
A tidy data frame.
}
\description{
Make a tidy data frame from smooth 3d distribution matrix
}
