% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{make_arg_grid}
\alias{make_arg_grid}
\title{Make variable grids for batch simulation}
\usage{
make_arg_grid(arg_set)
}
\arguments{
\item{arg_set}{An \code{arg_set} object. See \code{\link[=new_arg_set]{new_arg_set()}}
and \code{\link[=add_var]{add_var()}}.}
}
\value{
An \code{arg_grid} object.
}
\description{
This is the main function for making the variable grids.
}
\seealso{
\code{\link[=batch_simulation]{batch_simulation()}} for a concrete example.
}
