% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{modified_simulation}
\alias{modified_simulation}
\title{Modify a single simulation}
\usage{
modified_simulation(sim_fun, ele_list, default_list, bigmemory = TRUE, ...)
}
\arguments{
\item{ele_list}{An \code{ele_list} object generated by \code{\link[=fill_in_struct]{fill_in_struct()}}.}
}
\description{
Modify a single simulation
}
\keyword{internal}
