% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{narg}
\alias{narg}
\title{The number of arguments in an \code{arg_set}.}
\usage{
narg(arg_set)
}
\arguments{
\item{arg_set}{An \code{arg_set} object.}
}
\value{
An integer.
}
\description{
The number of arguments in an \code{arg_set}.
}
