% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{nele}
\alias{nele}
\title{The number of elements in an \code{arg_set}.}
\usage{
nele(arg_set)
}
\arguments{
\item{arg_set}{An \code{arg_set} object.}
}
\value{
An integer.
}
\description{
The number of elements in an \code{arg_set}.
}
