% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{npar}
\alias{npar}
\title{The number of parameters in a \code{var_set}.}
\usage{
npar(var_set)
}
\arguments{
\item{var_set}{A \code{var_set} object.}
}
\value{
An integer.
}
\description{
The number of parameters in a \code{var_set}.
}
