% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{plot.barrier}
\alias{plot.barrier}
\title{Plot the result of a \code{barrier} object}
\usage{
\method{plot}{barrier}(x, ...)
}
\arguments{
\item{x}{A `barrier` object.}

\item{...}{Not in use.}
}
\value{
The plot of the local minimums, the saddle point, and the lowest elevation path.
}
\description{
Plot the result of a \code{barrier} object
}
