% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_landscape.R
\name{plotly_ld}
\alias{plotly_ld}
\title{Plot landscape objects interactively}
\usage{
plotly_ld(object, ...)
}
\arguments{
\item{object}{A landscape object.}

\item{...}{Not in use.}
}
\value{
A plotly object.
}
\description{
Returns the plotly representation stored in a landscape object.
}
