% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{print.arg_grid}
\alias{print.arg_grid}
\title{Print an \code{arg_grid} object}
\usage{
\method{print}{arg_grid}(x, detail = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{detail}{Do you want to print the object details as a full list?}

\item{...}{Not in use.}
}
\value{
The printed result.
}
\description{
Print an \code{arg_grid} object
}
