% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{print.arg_set}
\alias{print.arg_set}
\title{Print an \code{arg_set} object.}
\usage{
\method{print}{arg_set}(x, detail = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{detail}{Do you want to print the object details as a full list?}

\item{...}{Not in use.}
}
\value{
The printed result.
}
\description{
Print an \code{arg_set} object.
}
