% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{print.batch_simulation}
\alias{print.batch_simulation}
\title{Print a batch_simulation}
\usage{
\method{print}{batch_simulation}(x, detail = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{detail}{Do you want to print the object details as a full list?}

\item{...}{Not in use.}
}
\description{
Print a batch_simulation
}
