% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conv.R
\name{print.check_conv}
\alias{print.check_conv}
\title{Print a `check_conv`}
\usage{
\method{print}{check_conv}(x, ask = TRUE, ...)
}
\arguments{
\item{x}{The object.}

\item{ask}{Ask to press enter to see the next plot?}

\item{...}{Not in use.}
}
\description{
Print a `check_conv`
}
