% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation_deprecated.R
\name{print.var_grid}
\alias{print.var_grid}
\title{Print a \code{var_grid} object}
\usage{
\method{print}{var_grid}(x, detail = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{detail}{Do you want to print the object details as a full list?}

\item{...}{Not in use.}
}
\value{
The printed result.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated. See \code{\link[=new_var_set]{new_var_set()}}.
}
\keyword{internal}
