% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{print.var_set}
\alias{print.var_set}
\title{Print a var_set}
\usage{
\method{print}{var_set}(x, detail = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{detail}{Do you want to print the object details as a full list?}

\item{...}{Not in use.}
}
\description{
Print a var_set
}
