% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_simulation_landscape.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{A function for reversed log transformation}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{The base of logarithm}
}
\description{
A function for reversed log transformation
}
