% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_landscape.R
\name{save_landscape}
\alias{save_landscape}
\title{Save landscape plots}
\usage{
save_landscape(l, path = NULL, selfcontained = FALSE, ...)
}
\arguments{
\item{l}{A landscape object}

\item{path}{The path to save the output. Default: "/pics/x_y.html".}

\item{selfcontained}{For 'plotly' plots, save the output as a self-contained html file? Default: FALSE.}

\item{...}{Other parameters passed to \code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}
or \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\value{
The function saves the plot to a specific path. It does not have a return value.
}
\description{
Save landscape plots
}
