% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{sim_fun_test2}
\alias{sim_fun_test2}
\title{A simple yet meaningful simulation function for testing}
\usage{
sim_fun_test2(...)
}
\arguments{
\item{...}{Parameters to be passed to \code{\link[=sim_fun_grad]{sim_fun_grad()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{sim_fun_test2()} was renamed to \code{sim_fun_grad()} to create a more
consistent API.
}
\keyword{internal}
