% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlandr-package.R
\docType{package}
\name{simlandr-package}
\alias{simlandr}
\alias{simlandr-package}
\title{simlandr: Simulation-Based Landscape Construction for Dynamical Systems}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for constructing potential landscapes for dynamical systems using Monte Carlo simulation. The method is based on the potential landscape definition by Wang et al. (2008) \doi{10.1073/pnas.0800579105} (also see Zhou & Li, 2016 \doi{10.1063/1.4943096} for further mathematical discussions) and can be used for a large variety of models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sciurus365.github.io/simlandr/}
  \item \url{https://github.com/Sciurus365/simlandr}
  \item Report bugs at \url{https://github.com/Sciurus365/simlandr/issues}
}

}
\author{
\strong{Maintainer}: Jingmeng Cui \email{jingmeng.cui@outlook.com} (\href{https://orcid.org/0000-0003-3421-8457}{ORCID})

Authors:
\itemize{
  \item Jingmeng Cui \email{jingmeng.cui@outlook.com} (\href{https://orcid.org/0000-0003-3421-8457}{ORCID})
}

}
\keyword{internal}
