% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{as.mcmc.list.list}
\alias{as.mcmc.list.list}
\title{Convert a list of simulation output to a mcmc.list object}
\usage{
\method{as.mcmc.list}{list}(x, ...)
}
\arguments{
\item{x}{A list of simulation output}

\item{...}{Not used}
}
\value{
A mcmc.list object
}
\description{
This function can be used to convert a list of simulation output to a mcmc.list object. This may be useful when the output of the simulation is a list of matrices, and you want to perform convergence checks using the functions in the coda package. See \code{\link[coda:mcmc.list]{coda::mcmc.list()}} for more information, and also see the examples in the documentation of \code{\link[=sim_SDE]{sim_SDE()}}.
}
