% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\name{as.mcmc.list}
\alias{as.mcmc.list}
\title{as.mcmc.list generic function}
\description{
See \code{\link[coda:as.mcmc.list]{coda::as.mcmc.list()}}, \code{\link[=as.mcmc.list.list]{as.mcmc.list.list()}} for details.
}
\keyword{internal}
