% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-autolayer.R
\name{autolayer}
\alias{autolayer}
\title{Autolayer generic function}
\description{
See \code{\link[ggplot2:autolayer]{ggplot2::autolayer()}}, \code{\link[=autolayer.barrier]{autolayer.barrier()}} for details.
}
\keyword{internal}
