% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{calculate_barrier}
\alias{calculate_barrier}
\title{Calculate barrier from a 2D distribution}
\usage{
calculate_barrier(
  d,
  start_location_value = c(0, 0),
  start_r = 0.1,
  end_location_value = c(0.7, 0.6),
  end_r = 0.15,
  base = exp(1)
)
}
\arguments{
\item{d}{A \code{3d_static_landscape} object or a \code{kde2d} distribution.}

\item{start_location_value, end_location_value}{The initial position (in value) for searching the start/end point.}

\item{start_r, end_r}{The searching radius for searching the start/end point.}

\item{base}{The base of the log function.}
}
\description{
Calculate barrier from a 2D distribution
}
