% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R, R/barrier_batch.R
\name{calculate_barrier}
\alias{calculate_barrier}
\alias{calculate_barrier.2d_density_landscape}
\alias{calculate_barrier.density}
\alias{calculate_barrier.3d_static_landscape}
\alias{calculate_barrier.list}
\alias{calculate_barrier.3d_animation_landscape}
\alias{calculate_barrier.3d_matrix_landscape}
\alias{calculate_barrier.2d_matrix_landscape}
\title{General function for calculating energy barrier}
\usage{
calculate_barrier(l, ...)

\method{calculate_barrier}{`2d_density_landscape`}(l, ...)

\method{calculate_barrier}{density}(l, ...)

\method{calculate_barrier}{`2d_density_landscape`}(l, ...)

\method{calculate_barrier}{`3d_static_landscape`}(l, ...)

\method{calculate_barrier}{list}(l, ...)

\method{calculate_barrier}{`3d_animation_landscape`}(l, ...)

\method{calculate_barrier}{`3d_matrix_landscape`}(l, ...)

\method{calculate_barrier}{`2d_matrix_landscape`}(l, ...)
}
\arguments{
\item{l}{A {landscape} or related project.}

\item{...}{Other parameters.}
}
\value{
A \code{barrier} object that contains the (batch) barrier calculation result(s).
}
\description{
General function for calculating energy barrier
}
\seealso{
\code{\link[=calculate_barrier_2d]{calculate_barrier_2d()}}, \code{\link[=calculate_barrier_2d_batch]{calculate_barrier_2d_batch()}}, \code{\link[=calculate_barrier_3d]{calculate_barrier_3d()}}, \code{\link[=calculate_barrier_3d_batch]{calculate_barrier_3d_batch()}}, \code{\link[=plot.barrier]{plot.barrier()}}
}
