% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{calculate_barrier_2d}
\alias{calculate_barrier_2d}
\title{Calculate barrier from a 2D landscape}
\usage{
calculate_barrier_2d(
  l,
  start_location_value = 0,
  start_r = 0.1,
  end_location_value = 0.7,
  end_r = 0.15,
  base = exp(1)
)
}
\arguments{
\item{l}{A \verb{2d_static_landscape} object (recommended) or a \code{density} distribution.}

\item{start_location_value, end_location_value}{The initial position (in value) for searching the start/end point.}

\item{start_r, end_r}{The searching radius for searching the start/end point.}

\item{base}{The base of the log function.}
}
\value{
A \code{barrier_2d} object that contains the barrier calculation result.
}
\description{
Calculate barrier from a 2D landscape
}
