% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conv.R
\name{check_conv}
\alias{check_conv}
\title{Check density convergence of simulation}
\usage{
check_conv(output, vars, sample_perc = 0.2, plot_type = "bin")
}
\arguments{
\item{output}{A matrix of simulation output.}

\item{vars}{The names of variables to check.}

\item{sample_perc}{The percentage of data sample for the initial, middle, and final stage of the simulation.}

\item{plot_type}{Which type of plots should be generated? ("bin" or "density")}
}
\value{
A \code{check_conv} object that contains the convergence checking result.
}
\description{
Check density convergence of simulation
}
