% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conv.R
\name{check_conv}
\alias{check_conv}
\alias{print.check_conv}
\title{Check density convergence of simulation}
\usage{
check_conv(output, vars, sample_perc = 0.2, plot_type = "bin")

\method{print}{check_conv}(x, ask = TRUE, ...)
}
\arguments{
\item{output}{A matrix of simulation output.}

\item{vars}{The names of variables to check.}

\item{sample_perc}{The percentage of data sample for the initial, middle, and final stage of the simulation.}

\item{plot_type}{Which type of plots should be generated? ("bin" or "density")}

\item{x}{The object.}

\item{ask}{Ask to press enter to see the next plot?}

\item{...}{Not in use.}
}
\value{
A \code{check_conv} object that contains the convergence checking result.
}
\description{
Check density convergence of simulation
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print a \code{check_conv} object.
}}

