% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{find_local_min_2d}
\alias{find_local_min_2d}
\title{Find local minimum of a 2d distribution}
\usage{
find_local_min_2d(dist, localmin, r)
}
\arguments{
\item{dist}{An \code{density} distribution object.}

\item{localmin}{Starting value of finding local minimum.}

\item{r}{Searching radius.}
}
\description{
Find local minimum of a 2d distribution
}
