% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{get_barrier_height}
\alias{get_barrier_height}
\title{Get the barrier height from a \code{barrier} object.}
\usage{
get_barrier_height(b)
}
\arguments{
\item{b}{A \code{barrier} object.}
}
\value{
A vector (for a single barrier calculation result) or a \code{data.frame} (for batch barrier calculation results) that contains the barrier heights on the landscape.
}
\description{
Get the barrier height from a \code{barrier} object.
}
