% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{get_geom}
\alias{get_geom}
\title{Get a ggplot2 geom layer that can be added to a ggplot2 landscape plot}
\usage{
get_geom(b, path = TRUE)
}
\arguments{
\item{b}{A \code{barrier} object.}

\item{path}{Show the minimum energy path in the graph?}
}
\value{
A \code{ggplot2} geom (formally a \code{LayerInstance} object) that can be added to an existing \code{ggplot}.
}
\description{
This layer can show the saddle point (2d) and the minimal energy path (3d) on the landscape.
}
