% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{get_geom}
\alias{get_geom}
\title{Get a ggplot2 layer from a barrier object}
\usage{
get_geom(object, ...)
}
\arguments{
\item{object}{A \code{barrier} object.}

\item{...}{Not in use.}
}
\value{
A \code{ggplot2} layer that can be added to an existing landscape plot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated. Use \code{\link[=autolayer]{autolayer()}} instead. (See \code{\link[=autolayer.barrier]{autolayer.barrier()}}.)
}
\keyword{internal}
