% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_simulation_landscape.R
\name{make_2d_density}
\alias{make_2d_density}
\title{Make 2D density-based landscape plot for a single simulation output}
\usage{
make_2d_density(output, x, adjust = 50, from = -0.1, to = 1, Umax = 5)
}
\arguments{
\item{output}{A matrix of simulation output.}

\item{x}{The name of the target variable.}

\item{adjust, from, to}{Passed to \code{density}.}

\item{Umax}{The maximum displayed value of potential.}
}
\value{
A \code{2d_density_landscape} object.
}
\description{
Make 2D density-based landscape plot for a single simulation output
}
