% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_simulation_landscape.R
\name{make_2d_matrix}
\alias{make_2d_matrix}
\title{Make a matrix of 2d graphs for two parameters}
\usage{
make_2d_matrix(bs, x, rows, cols, adjust = 50, from = -0.1, to = 1, zmax = 5)
}
\arguments{
\item{bs}{A \code{batch_simulation} object created by \code{\link{batch_simulation}.}}

\item{x, rows, cols}{The names of the target variables.}

\item{adjust, from, to}{Passed to \code{density}.}

\item{zmax}{The maximum displayed value of potential.}
}
\description{
Make a matrix of 2d graphs for two parameters
}
