% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_simulation_landscape.R
\name{make_2d_matrix}
\alias{make_2d_matrix}
\title{Make a matrix of 2d graphs for two parameters}
\usage{
make_2d_matrix(
  bs,
  x,
  rows = NULL,
  cols,
  adjust = 50,
  from = -0.1,
  to = 1,
  Umax = 5,
  individual_landscape = FALSE
)
}
\arguments{
\item{bs}{A \code{batch_simulation} object created by \verb{[batch_simulation].}}

\item{x, rows, cols}{The names of the target variables.
If \code{rows} is \code{NULL}, only a vector of graphs will be generated.}

\item{adjust, from, to}{Passed to \code{density}.}

\item{Umax}{The maximum displayed value of potential.}

\item{individual_landscape}{Make individual landscape for each simulation?}
}
\value{
A \verb{2d_matrix_landscape} object that describes the landscape of the system, including the smoothed distribution and the landscape plot.
}
\description{
Make a matrix of 2d graphs for two parameters
}
