% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_simulation_landscape.R
\name{make_2d_static}
\alias{make_2d_static}
\title{Make 2D static landscape plot for a single simulation output}
\usage{
make_2d_static(output, x, adjust = 50, from = -0.1, to = 1, Umax = 5)
}
\arguments{
\item{output}{A matrix of simulation output.}

\item{x}{The name of the target variable.}

\item{adjust, from, to}{Passed to \code{density}.}

\item{Umax}{The maximum displayed value of potential.}
}
\value{
A \verb{2d_static_landscape} object that describes the landscape of the system, including the smooth distribution and the landscape plot.
}
\description{
Make 2D static landscape plot for a single simulation output
}
