% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_simulation_landscape.R
\name{make_3d_animation}
\alias{make_3d_animation}
\title{Make 3d animations from multiple simulations}
\usage{
make_3d_animation(
  bs,
  x,
  y,
  fr,
  Umax = 5,
  n = 200,
  lims = c(-0.1, 1.1, -0.1, 1.1),
  h = 0.001,
  kde_fun = "ks",
  individual_landscape = FALSE,
  mat_3d = TRUE
)
}
\arguments{
\item{bs}{A \code{batch_simulation} object created by \verb{[batch_simulation].}}

\item{x, y, fr}{The names of the target variables.
\code{fr} corresponds to the \code{frame} parameter in 'plotly'.}

\item{Umax}{The maximum displayed value of potential.}

\item{n, lims, h, kde_fun}{Passed to \code{make_2d_kernel_dist}}

\item{individual_landscape}{Make individual landscape for each simulation?}

\item{mat_3d}{Also make heatmap matrix?}
}
\value{
A \verb{3d_animation_landscape} object that describes the landscape of the system, including the smoothed distribution and the landscape plot.
}
\description{
Make 3d animations from multiple simulations
}
