% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_simulation_landscape.R
\name{make_3d_animation_multisim}
\alias{make_3d_animation_multisim}
\title{Make 3d animations from multiple simulations}
\usage{
make_3d_animation_multisim(bs, x, y, fr, zmax = 5, individual_plot = FALSE)
}
\arguments{
\item{bs}{A \code{batch_simulation} object created by \code{\link{batch_simulation}.}}

\item{x, y, fr}{The names of the target variables.
\code{fr} corresponds to the \code{frame} parameter in plotly.}

\item{zmax}{The maximum displayed value of potential.}

\item{individual_plot}{Make individual plot for each var value?}
}
\description{
Make 3d animations from multiple simulations
}
