% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_simulation_landscape.R
\name{make_3d_animation_multisim}
\alias{make_3d_animation_multisim}
\title{Make 3d animations from multiple simulations}
\usage{
make_3d_animation_multisim(
  bs,
  x,
  y,
  fr,
  zmax = 5,
  n = 200,
  lims = c(-0.1, 1.1, -0.1, 1.1),
  h = 0.001,
  kde_fun = "ks",
  individual_plot = FALSE,
  mat_3d = TRUE
)
}
\arguments{
\item{bs}{A \code{batch_simulation} object created by \code{\link{batch_simulation}.}}

\item{x, y, fr}{The names of the target variables.
\code{fr} corresponds to the \code{frame} parameter in plotly.}

\item{zmax}{The maximum displayed value of potential.}

\item{n, lims, h, kde_fun}{Passed to \code{make_kernel_dist}}

\item{individual_plot}{Make individual plot for each var value?}

\item{mat_3d}{Also make heatmap matrix?}
}
\description{
Make 3d animations from multiple simulations
}
