% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_simulation_landscape.R
\name{make_3d_matrix}
\alias{make_3d_matrix}
\title{Make a matrix or vector of 3d heatmap graphs for two parameters}
\usage{
make_3d_matrix(
  bs,
  x,
  y,
  rows = NULL,
  cols,
  Umax = 5,
  n = 200,
  lims = c(-0.1, 1.1, -0.1, 1.1),
  h = 0.001,
  kde_fun = "ks",
  individual_landscape = FALSE
)
}
\arguments{
\item{bs}{A \code{batch_simulation} object created by \verb{[batch_simulation].}}

\item{x, y, rows, cols}{The names of the target variables.
If \code{rows} is \code{NULL}, only a vector of graphs will be generated.}

\item{Umax}{The maximum displayed value of potential.}

\item{n, lims, h}{Passed to \code{\link[ks:kde]{ks::kde()}} or \code{\link[MASS:kde2d]{MASS::kde2d()}}.
If using \code{ks::kde}, \code{H = diag(h, 2, 2)}.
Note: the definition of bandwidth (\code{h}) is different in two functions.
To get a similar output, the \code{h} is about 50 to 5000 times smaller for \code{\link[ks:kde]{ks::kde()}} than \code{\link[MASS:kde2d]{MASS::kde2d()}}}

\item{kde_fun}{Which to use? Choices: "ks" \code{ks::kde} (default; faster and taking less memory); "MASS" \code{MASS::kde2d}.}

\item{individual_landscape}{Make individual landscape for each simulation?}
}
\value{
A \verb{3d_matrix_landscape} object that describes the landscape of the system, including the smoothed distribution and the landscape plot.
}
\description{
(Note: a matrix of interactive maps is currently not supported.)
}
