% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_simulation_landscape.R
\name{make_3d_matrix}
\alias{make_3d_matrix}
\title{Make a matrix or vector of 3d heatmap graphs for two parameters}
\usage{
make_3d_matrix(
  bs,
  x,
  y,
  rows = NULL,
  cols,
  Umax = 5,
  n = 200,
  lims = c(-0.1, 1.1, -0.1, 1.1),
  h = 0.001,
  kde_fun = "ks",
  individual_landscape = FALSE
)
}
\arguments{
\item{bs}{A \code{batch_simulation} object created by \code{\link{batch_simulation}.}}

\item{x, y, rows, cols}{The names of the target variables.
If `rows` is `NULL`, only a vector of graphs will be generated.}

\item{Umax}{The maximum displayed value of potential.}

\item{n, lims, h, kde_fun}{Passed to \code{\link{make_2d_kernel_dist}}}

\item{individual_landscape}{Make individual landscape for each simulation?}
}
\description{
(Note: a matrix of interactive maps is currently not supported.)
}
