% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_simulation_landscape.R
\name{make_3d_static}
\alias{make_3d_static}
\title{Make 3D static landscape plots from simulation output}
\usage{
make_3d_static(
  output,
  x,
  y,
  zmax = 5,
  save = FALSE,
  path = NULL,
  selfcontained = F,
  n = 200,
  lims = c(-0.1, 1.1, -0.1, 1.1),
  h = 0.1
)
}
\arguments{
\item{output}{A matrix of simulation output.}

\item{x, y}{The name of the target variable.}

\item{zmax}{The maximum displayed value of potential.}

\item{save}{Save the output as a html file? Default: FALSE.}

\item{path}{The path to save the output. Default: "/pics/x_y.html".}

\item{selfcontained}{Save the output as a selfcontained html file? Default: FALSE.}

\item{n, lims, h}{Passed to \code{\link[MASS]{kde2d}}}
}
\value{
A \code{3d_static_landscape}, \code{landscape} object.
}
\description{
Make 3D static landscape plots from simulation output
}
