% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_simulation_landscape.R
\name{make_3d_static}
\alias{make_3d_static}
\title{Make 3D static landscape plots from simulation output}
\usage{
make_3d_static(
  output,
  x,
  y,
  Umax = 5,
  n = 200,
  lims = c(-0.1, 1.1, -0.1, 1.1),
  h = 0.001,
  kde_fun = "ks"
)
}
\arguments{
\item{output}{A matrix of simulation output.}

\item{x, y}{The name of the target variable.}

\item{Umax}{The maximum displayed value of potential.}

\item{n, lims, h}{Passed to \code{\link[ks:kde]{ks::kde()}} or \code{\link[MASS:kde2d]{MASS::kde2d()}}.
If using \code{ks::kde}, \code{H = diag(h, 2, 2)}.
Note: the definition of bandwidth (\code{h}) is different in two functions.
To get a similar output, the \code{h} is about 50 to 5000 times smaller for \code{\link[ks:kde]{ks::kde()}} than \code{\link[MASS:kde2d]{MASS::kde2d()}}}

\item{kde_fun}{Which to use? Choices: "ks" \code{ks::kde} (default; faster and taking less memory); "MASS" \code{MASS::kde2d}.}
}
\value{
A \verb{3d_static_landscape} object that describes the landscape of the system, including the smooth distribution and the landscape plot.
}
\description{
Make 3D static landscape plots from simulation output
}
