% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_simulation_landscape.R
\name{make_3d_static}
\alias{make_3d_static}
\title{Make 3D static landscape plots from simulation output}
\usage{
make_3d_static(
  output,
  x,
  y,
  Umax = 5,
  n = 200,
  lims = c(-0.1, 1.1, -0.1, 1.1),
  h = 0.001,
  kde_fun = "ks"
)
}
\arguments{
\item{output}{A matrix of simulation output.}

\item{x, y}{The name of the target variable.}

\item{Umax}{The maximum displayed value of potential.}

\item{n, lims, h, kde_fun}{Passed to \code{\link{make_2d_kernel_dist}}}
}
\value{
A \code{3d_static_landscape}, \code{landscape} object.
}
\description{
Make 3D static landscape plots from simulation output
}
