% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_simulation_landscape.R
\name{make_4d_static}
\alias{make_4d_static}
\title{Make 4D static space-color plots from simulation output}
\usage{
make_4d_static(
  output,
  x,
  y,
  z,
  Umax = 5,
  n = 50,
  lims = c(-0.1, 1.1, -0.1, 1.1, -0.1, 1.1),
  h = 0.001
)
}
\arguments{
\item{output}{A matrix of simulation output.}

\item{x, y, z}{The name of the target variable.}

\item{Umax}{The maximum displayed value of potential.}

\item{n, lims, h}{Passed to \code{\link{make_3d_kernel_dist}}}
}
\value{
A \code{4d_static_landscape}, \code{landscape} object.
}
\description{
Make 4D static space-color plots from simulation output
}
