% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier_batch.R
\name{make_barrier_grid_2d}
\alias{make_barrier_grid_2d}
\title{Make a grid for calculating barriers for 2d landscapes}
\usage{
make_barrier_grid_2d(
  ag,
  start_location_value,
  start_r,
  end_location_value,
  end_r,
  df = NULL,
  print_template = FALSE
)
}
\arguments{
\item{ag}{An \code{arg_grid} object.}

\item{start_location_value, start_r, end_location_value, end_r}{Default values for finding local minimum. See \code{\link[=calculate_barrier]{calculate_barrier()}}.}

\item{df}{A data frame for the variables. Use \code{print_template = TRUE} to get a template.}

\item{print_template}{Print a template for \code{df}.}
}
\value{
A \code{barrier_grid_2d} object that specifies the condition for each barrier calculation.
}
\description{
Make a grid for calculating barriers for 2d landscapes
}
