% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier_batch.R
\name{make_barrier_grid_3d}
\alias{make_barrier_grid_3d}
\title{Make a grid for calculating barriers for 3d landscapes}
\usage{
make_barrier_grid_3d(
  vg,
  start_location_value = c(0, 0),
  start_r = 0.1,
  end_location_value = c(0.7, 0.6),
  end_r = 0.15,
  df = NULL,
  print_template = FALSE
)
}
\arguments{
\item{vg}{A \code{var_grid} object.}

\item{start_location_value, start_r, end_location_value, end_r}{Default values for finding local minimum. See \code{\link{calculate_barrier_3d_batch}}.}

\item{df}{A data frame for the variables. Use \code{print_template = TRUE} to get a template.}

\item{print_template}{Print a template for \code{df}.}
}
\description{
Make a grid for calculating barriers for 3d landscapes
}
