% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{make_var_grid}
\alias{make_var_grid}
\title{Make variable grids for batch simulation}
\usage{
make_var_grid(var_set)
}
\arguments{
\item{var_set}{A \code{var_set} object. See \code{\link{new_var_set}}
and \code{\link{add_var}}.}
}
\value{
A \code{var_grid} object.
}
\description{
This is the main function for making the variable grids.
}
\seealso{
\code{\link{batch_simulation}} for a concrete example.
}
