% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation_deprecated.R
\name{npar}
\alias{npar}
\title{The number of parameters in a \code{var_set}.}
\usage{
npar(var_set)
}
\arguments{
\item{var_set}{A \code{var_set} object.}
}
\value{
An integer.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated. See \code{\link[=new_var_set]{new_var_set()}}.
}
