% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_simulation.R
\name{nvar}
\alias{nvar}
\title{The number of variables in a \code{var_set}.}
\usage{
nvar(var_set)
}
\arguments{
\item{var_set}{A \code{var_set} object.}
}
\value{
An integer.
}
\description{
The number of variables in a \code{var_set}.
}
