% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_plot.R
\name{plot.landscape}
\alias{plot.landscape}
\title{Make plots from landscape objects}
\usage{
\method{plot}{landscape}(x, ...)
}
\arguments{
\item{x}{A landscape object}

\item{...}{Not in use.}
}
\description{
Make plots from landscape objects
}
