% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_landscape.R
\name{plot.landscape}
\alias{plot.landscape}
\title{Make plots from landscape objects}
\usage{
\method{plot}{landscape}(x, index = 1, ...)
}
\arguments{
\item{x}{A landscape object}

\item{index}{Default is 1. For some landscape objects, there is a second plot (usually 2d heatmaps for 3d landscapes)
or a third plot (usually 3d matrices for 3d animations).
Use \code{index = 2} to plot that one.}

\item{...}{Not in use.}
}
\value{
The plot.
}
\description{
Make plots from landscape objects
}
