% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{sim_fun_test}
\alias{sim_fun_test}
\title{A simple simulation function for testing}
\usage{
sim_fun_test(arg1, arg2, length = 1000)
}
\arguments{
\item{arg1, arg2}{Two parameters. \code{arg1} contains \code{ele1};
\code{arg2} contains \code{ele2} and \code{ele3}.}

\item{length}{The length of simulation.}
}
\value{
A matrix of simulation results.
}
\description{
A simple simulation function for testing
}
\seealso{
\code{\link[=sim_fun_grad]{sim_fun_grad()}} and \code{\link[=sim_fun_nongrad]{sim_fun_nongrad()}} for more realistic
examples.
}
