% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{sim_fun_test}
\alias{sim_fun_test}
\title{A simple simulation function for testing}
\usage{
sim_fun_test(par1, par2, length = 1000)
}
\arguments{
\item{par1, par2}{Two parameters. \code{par1} contains \code{var1};
\code{par2} contains \code{var2} and \code{var3}.}

\item{length}{The length of simulation.}
}
\value{
A matrix of simulation results.
}
\description{
A simple simulation function for testing
}
\seealso{
\code{\link{sim_fun_test2}} for a more realistic example. \code{\link{batch_simulation}} for a concrete example.
}
