% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{sim_fun_test2}
\alias{sim_fun_test2}
\title{A simple yet meaningful simulation function for testing}
\usage{
sim_fun_test2(
  par1 = list(x = 1, y = 1),
  par2 = list(sigmasq = 25, c = 1),
  length = 1e+05
)
}
\arguments{
\item{par1, par2}{Two sets of parameters. \code{par1} contains \code{x} and \code{y};
\code{par2} contains \code{c}, which controls the shape of the landscape, and \code{sigmasq}, which controls the amplitude of noise.}

\item{length}{The length of simulation.}
}
\value{
A matrix of simulation results.
}
\description{
This is a toy system with bi-stability. You can use it to try out functions.
Model specification:
\deqn{dx/dt = cy - y^3 + \sigma dW}
\deqn{y = x + \sigma dW}
}
\seealso{
\code{\link{sim_fun_test}} and \code{\link{batch_simulation}}.
}
