% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barrier.R
\name{summary.barrier}
\alias{summary.barrier}
\title{Summarize the barrier height from a \code{barrier} object.}
\usage{
\method{summary}{barrier}(object, ...)
}
\arguments{
\item{object}{A \code{barrier} object.}

\item{...}{Not in use.}
}
\value{
A vector (for a single barrier calculation result) or a \code{data.frame} (for batch barrier calculation results) that contains the barrier heights on the landscape.
}
\description{
Summarize the barrier height from a \code{barrier} object.
}
